// src/internal/dashboard/models/ShopServices.tsx
export type CustomerServices = {
    ID: number
    Created_at?: Date
    Updated_at?: Date
    title: string
    description: string
    price: number
    service_duration: number
    category: string
    created_by_id: number
    rating: number
    reviews_count: number
    preview_url: string
}